Introduce dir abstraction
With the goal of adding playback of non-local files, this change introduces a directory abstraction. The directory is the source of the music data and supports reading files (steady-state) and iterating (scanning) them.

This implies two changes to the existing codebase: 1) removal of the assumption that one can read files as local files, instead needing to read them through the abstraction, and 2) removal of the usage of ‘full path’s as they no longer make sense as the ‘full path’ could be a URL or other.